requirements.txt
langchain_openai
langchain_core
python-dotenv
streamlit
langchain_community
pypdf
chromadb
langchain_Ollama
langchain_groq
